Fix 32-bit overflow in lineardb3 offset calculation preventing files > 4GB
When calculating the byte offset for a record (`fileIndex * recordSizeBytes`), the multiplication was performed using 32-bit integers. This caused an arithmetic overflow once the file reached ~4.3 GB (exactly 178,956,971 records for mapTime.db).

Because the resulting overflowed offset did not match the end of the file, lineardb3's safety check correctly aborted the write. This prevented file corruption but caused the database to reject all new records, leaving new time-based items to freeze on the map once they fell out of the in memory cache.

Fixed by casting to `uint64_t` prior to multiplication in lineardb3.cpp.

+++ server/lineardb3.cpp
@@ -1020,8 +1020,8 @@ static int LINEARDB3_considerFingerprintBucket( LINEARDB3 *inDB,
             
         uint64_t filePosRec = 
             LINEARDB3_HEADER_SIZE +
-            inBucket->fileIndex[ i ] * 
-            inDB->recordSizeBytes;
+            (uint64_t)( inBucket->fileIndex[ i ] ) * 
+            (uint64_t)( inDB->recordSizeBytes );
             
         if( !emptyRec ) {
             
@@ -1227,8 +1227,8 @@ int LINEARDB3_getOrPut( LINEARDB3 *inDB, const void *inKey, void *inOutValue,
 
             uint64_t filePosRec = 
                 LINEARDB3_HEADER_SIZE +
-                newBucket->fileIndex[0] * 
-                inDB->recordSizeBytes;
+                (uint64_t)( newBucket->fileIndex[0] ) * 
+                (uint64_t)( inDB->recordSizeBytes );
 
             // don't seek unless we have to
             if( inDB->lastOp == opRead ||
@@ -1322,7 +1322,7 @@ int LINEARDB3_Iterator_next( LINEARDB3_Iterator *inDBi,
         
         uint64_t fileRecPos = 
             LINEARDB3_HEADER_SIZE + 
-            inDBi->nextRecordIndex * db->recordSizeBytes;
+            (uint64_t)( inDBi->nextRecordIndex ) * (uint64_t)( db->recordSizeBytes );
         
                     
         if( db->lastOp == opWrite ||
